requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
pillow>=10.0.0
pydantic>=2.6.0
